Drop sqlite support
Problem: sqlite does not support many feature normal databases do. It
is hard to write migrations for sqlite and it does not support NULLS
FIRST in the ORDER BY, which is needed for deterministic ordering of
bookmarks.

Solution: I drop sqlite support from now on and adapt tests to use
postgres database.